Move BuildIntraEpitopesCDS into pipeline_utils and setup v2
